better organize test helper files
