+++ src/components/createTestCase/CreateTestCase.test.tsx
@@ -24,7 +24,7 @@ import { ExecutionResult } from "fqm-execution/build/types/Calculator";
 import {
   buildMeasureBundle,
   successfulExecutionResults,
-} from "./ExectionTestHelpers";
+} from "../../util/CalculationTestHelpers";
 
 //temporary solution (after jest updated to version 27) for error: thrown: "Exceeded timeout of 5000 ms for a test.
 jest.setTimeout(60000);
